feat(frontend): Instrument http handlers
OpenTelemetry is is made of exporters, the easier to use is the stdout
one. It prints JSON to the process stdout.

Stdout is a good exporter but not the one you should use in production.
There are a lot of open source tracer around: Zipkin, Jaeger, Honeycomb,
AWS X-Ray, Google StackDriver. I tend to use Jaeger because it is in Go
and it is open source.

This commit adds the flag `--tracer` by default it is set to stdout, but
if you use `--tracer jaeger` the traces will be send to Jaeger. You can
override the Jaeger URl with `--tracer-jaeger-address`

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>